requirements.txt
numpy==1.24.3
Flask==2.3.2
inference==0.10.0
opencv-python==4.8.0.76
pillow==10.3.0